refactor: change `percentage_progress` type in `SSavingGoalProgress` from int to float

+++ app/schemas/saving_goals_schemas.py
@@ -109,7 +109,7 @@ class SSavingGoalProgress(BaseModel):
     current_amount: int
     target_amount: int
     rest_amount: int
-    percentage_progress: int
+    percentage_progress: float
     days_left: int
     expected_daily_payment: int
